+++ back/controllers/ctrlPost.js
@@ -1,4 +1,5 @@
 const Post = require("../models/modelPost")
+const fs = require('fs');
 
 //Afficher tout les posts
 
@@ -16,7 +17,11 @@ exports.getAllPosts = async (req, res, next) => {
 
 exports.createPost = async (req, res, next) => {
     try {
-        const newPost = { ...req.body, user_id: req.auth.user_id };
+        const newPost = { 
+            ...req.body, 
+            user_id: req.auth.user_id,  
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` 
+        };
         const post = await Post.create(newPost);
         res.status(201).json({ message: 'Post created', post });
     } catch (error) {
@@ -44,7 +49,15 @@ exports.getOnePost = async (req, res, next) => {
 exports.updatePost = async (req, res, next) => {
     try {
         const postId = req.params.id;  
-        const postObject = req.body;
+        const postObject = req.file ? {
+            ...req.body,
+            user_id: req.auth.user_id,
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        } : {
+            ...req.body,
+            user_id: req.auth.user_id,
+        }
+        
         delete postObject.id; 
 
         // Mettre à jour l'utilisateur
@@ -61,14 +74,26 @@ exports.updatePost = async (req, res, next) => {
 
 exports.deletePost = async (req, res, next) => {
     try {
-        
         const postId = req.params.id; 
-
-        await Post.destroy({where: { id: postId }});
-        res.status(200).json({ message: 'Post deleted!' });
-
+        const post = await Post.findByPk(postId)
+        if (!post) {
+            return res.status(404).json({ message: 'Post not found!' });
         }
-     catch (error) {
+        if (post.imageUrl) {
+            const filename = post.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, async (err) => {
+                if (err) {
+                    console.error('Error deleting image:', err);
+                }
+                await Post.destroy({where: { id: postId }});
+                res.status(200).json({ message: 'Post deleted!' });
+            });
+        } else {
+            await Post.destroy({where: { id: postId }});
+            res.status(200).json({ message: 'Post deleted!' });
+        }
+    }
+    catch (error) {
         res.status(500).json({ error: error.message });
     }
 };
